+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v3.0.2 (06-02-2019)
+
+New strategy for client side
+
 ## v3.0.1 (05-02-2019)
 
 Fix for when the method throws
@@ -12,7 +16,9 @@ Modify the client side so we can also hook into method calls that exist only on 
 
 **Breaking change** The hooks on the client side are now triggered before and after the server call. Not before and after the simulated client-side call.
 
-**Breaking change** Even when the method throws, the after hooks are still called. `this._error` will contain the error object that was thrown.
+**Breaking change** in the hooks running after the method, the result is now accessible via `this.result` instead of `this._result`. Changing `this.result` will change the passed result.
+
+**Breaking change** Even when the method throws, the after hooks are still called. `this.error` will contain the error object that was thrown.
 
 ## v2.0.0 (12-03-2018)